refactor(web): extract TrainingStudio pure helpers + split panels (sc-4199)
F-WEB-20 follow-on to the JoyCaption corpus extraction (#552). The 2.1k-line
TrainingStudio screen mixed pure data/payload builders with the UI. Extract the
pure helpers into dedicated, testable src/training/ modules and split the two
large panels into their own components — behavior is unchanged.

- src/training/drafts.js: draft<->value/number helpers (asText, numericDraft,
  numberFromDraft, boundedNumber, integerFromDraft, compactObject,
  normalizeTrainingAdapterVersion).
- src/training/datasetHelpers.js: dataset summaries, selection-key derivation,
  owned-asset/asset-id normalizers, datasetHealth, datasetPayload, etc.
- src/training/trainingConfig.js: target/preset label maps + option arrays and
  the two buried builders configDraftFromTarget and trainingConfigSnapshot, plus
  configValidation/presetsForTarget/defaultPresetForTarget/samplePromptsFromTrigger.
- src/training/trainingConfig.test.js: 13 new unit tests for configDraftFromTarget
  and trainingConfigSnapshot (the finding named these "testable but buried").
- src/screens/training/DatasetEditorPanel.jsx and ConfigureJobPanel.jsx: the
  dataset-editor and configure-job panels, with the screen owning all state and
  handlers and passing them in as props.

TrainingStudio.jsx re-exports configDraftFromTarget/trainingConfigSnapshot so the
existing screen test keeps importing them from the screen.

Verified in apps/web: npm run lint (0 errors), npm run test (396 pass), npm run build.
